feat(cli): extract and modularize dlock CLI and logic for better structure
- Moved core logic from monolithic 'main.go' into organized 'pkg/dlock' package
- Introduced 'cmd/dlock/main.go' as the new CLI entry point
- Added Makefile for streamlined build, install, test, and multi-platform compilation
- Refactored ADB interaction, methods, validation, and types into separate files
- Added usage example in 'examples/usage_example.go'
- Updated GitHub workflow to build using new 'cmd/dlock' entry point
