+++ .github/workflows/main.yaml
@@ -22,19 +22,19 @@ jobs:
         run: |
           go get
           echo "Preparing Windows binary"
-          GOOS=windows GOARCH=amd64 go build -ldflags="-X 'main.AppVersion=${{ github.ref_name}} '" -o dlock.exe .
+          GOOS=windows GOARCH=amd64 go build -ldflags="-X 'main.AppVersion=${{ github.ref_name}} '" -o dlock.exe ./cmd/dlock
           zip -j dlock-win.zip dlock.exe
           rm dlock.exe
           echo "Preparing macOS amd64 binary"
-          GOOS=darwin GOARCH=amd64 go build -ldflags="-X 'main.AppVersion=${{ github.ref_name}} '"  -o dlock .
+          GOOS=darwin GOARCH=amd64 go build -ldflags="-X 'main.AppVersion=${{ github.ref_name}} '"  -o dlock ./cmd/dlock
           zip -j dlock-mac-amd64.zip dlock
           rm dlock
           echo "Preparing macOS arm64 binary"
-          GOOS=darwin GOARCH=arm64 go build -ldflags="-X 'main.AppVersion=${{ github.ref_name}} '"  -o dlock .
+          GOOS=darwin GOARCH=arm64 go build -ldflags="-X 'main.AppVersion=${{ github.ref_name}} '"  -o dlock ./cmd/dlock
           zip -j dlock-mac-arm64.zip dlock
           rm dlock
           echo "Preparing Linux binary"
-          GOOS=linux GOARCH=amd64 go build -ldflags="-X 'main.AppVersion=${{ github.ref_name}} '"  -o dlock .
+          GOOS=linux GOARCH=amd64 go build -ldflags="-X 'main.AppVersion=${{ github.ref_name}} '"  -o dlock ./cmd/dlock
           zip -j dlock-linux.zip dlock
           rm dlock
